PHP 8.5 compat: Moved from "__sleep/__wakeup" to "__serialize/__unserialize", Removed non-canonical cast names

+++ src/system.php
@@ -1178,7 +1178,7 @@ function generatePW($len = 8, $case_sensitive=true, $chars='')
     }
     else
     {
-        mt_srand ((double) microtime(false) * 1000000);
+        mt_srand ((float) microtime(false) * 1000000);
         while( strlen($res) < $len )
             $res .= $chars[mt_rand(0,strlen($chars)-1)];
     }
